+++ .dart_tool/extension_discovery/vs_code.json
@@ -1,0 +1,1 @@
+{"version":2,"entries":[{"package":"theparkingdeals","rootUri":"../","packageUri":"lib/"}]}
