+++ .dart_tool/package_config.json
@@ -104,6 +104,12 @@
       "languageVersion": "2.12"
     },
     {
+      "name": "intl",
+      "rootUri": "file:///Users/sube/.pub-cache/hosted/pub.dev/intl-0.17.0",
+      "packageUri": "lib/",
+      "languageVersion": "2.12"
+    },
+    {
       "name": "leak_tracker",
       "rootUri": "file:///Users/sube/.pub-cache/hosted/pub.dev/leak_tracker-10.0.5",
       "packageUri": "lib/",
@@ -314,7 +320,7 @@
       "languageVersion": "3.5"
     }
   ],
-  "generated": "2024-10-13T18:17:24.010613Z",
+  "generated": "2024-10-15T11:52:59.937218Z",
   "generator": "pub",
   "generatorVersion": "3.5.3",
   "flutterRoot": "file:///Users/sube/development/flutter",
