+++ .dart_tool/package_config_subset
@@ -58,6 +58,10 @@ http_parser
 2.12
 file:///Users/sube/.pub-cache/hosted/pub.dev/http_parser-4.0.2/
 file:///Users/sube/.pub-cache/hosted/pub.dev/http_parser-4.0.2/lib/
+intl
+2.12
+file:///Users/sube/.pub-cache/hosted/pub.dev/intl-0.17.0/
+file:///Users/sube/.pub-cache/hosted/pub.dev/intl-0.17.0/lib/
 leak_tracker
 3.2
 file:///Users/sube/.pub-cache/hosted/pub.dev/leak_tracker-10.0.5/
