+++ lib/modules/auth/signin/signin.dart
@@ -165,7 +165,7 @@ class _SignInScreenState extends State<SignInScreen> {
                     child: CustomButton(
                       label: "Sign In",
                       onPressed: () {
-                         Navigator.pushNamed(context, AppRoutes.myBookingExpand);
+                         Navigator.pushNamed(context, AppRoutes.myBooking);
                       },
                       backgroundColor: background,
                       fontSize: 20,
