+++ lib/routes/route.dart
@@ -1,8 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:theparkingdeals/modules/auth/signin/signin.dart';
 import 'package:theparkingdeals/modules/auth/signup/signup.dart';
-import 'package:theparkingdeals/modules/my_booking/my_booking_screen.dart';
-import 'package:theparkingdeals/modules/my_profile/my_profile_screen.dart';
+import 'package:theparkingdeals/modules/get_quote/presenters/quote_screen.dart';
+import 'package:theparkingdeals/modules/my_booking/presenters/my_booking_expand_screen.dart';
+import 'package:theparkingdeals/modules/my_booking/presenters/my_booking_screen.dart';
+import 'package:theparkingdeals/modules/my_profile/presenters/my_profile_screen.dart';
 import 'package:theparkingdeals/modules/splash_screen/splash_screen.dart';
 
 class AppRoutes {
@@ -12,6 +14,7 @@ class AppRoutes {
   static const String myProfile = '/myProfile';
   static const String myBooking = '/myBooking';
   static const String myBookingExpand = '/myBookingExpand';
+  static const String getQuote = '/getQuote';
 
   static Route<dynamic> generateRoute(RouteSettings settings) {
     switch (settings.name) {
@@ -26,7 +29,9 @@ class AppRoutes {
       case myBooking:
         return MaterialPageRoute(builder: (_) => const MyBookingScreen());
       case myBookingExpand:
-        return MaterialPageRoute(builder: (_) => const MyBookingScreen());
+        return MaterialPageRoute(builder: (_) => const MyBookingExpandScreen());
+      case getQuote:
+        return MaterialPageRoute(builder: (_) => const QuoteScreen());
       default:
         return MaterialPageRoute(
             builder: (_) => Scaffold(
