+++ pubspec.yaml
@@ -12,6 +12,7 @@ dependencies:
     sdk: flutter
   flutter_svg: ^2.0.10+1
   google_fonts: ^6.2.1
+  intl: ^0.17.0
 
   cupertino_icons: ^1.0.8
